requirements.txt
numpy
pandas
matplotlib
markdown
jinja2
django
pygments
django-background-tasks
